[#1984] restructure installation to remove rodsBoot user
This required triggering on whether the rodsBoot user exists.

If the rodsBoot account exists during installation, this is the
first run of irods_setup.pl:
 - Setup the database and admin account
 - Bounce the server, come up in admin account
 - Delete the rodsBoot user

Otherwise, if the rodsBoot account does not exist, this is not
the first run of irods_setup.pl:
 - Bounce the server, come up in admin account
 - Run largely redundant setup

This keeps rodsBoot from being an extra account lying around on
production installations.

This also allows re-installation of a package atop an existing
working iCAT without error.