[#2009] [#2019] update mechanism for catalog schema revisions
This reads from a directory of update files (schema_updates)
and walks them in order.  The current schema version is found
from the existing database and the target schema version is
found from the top level VERSION file.

This mechanism should upgrade cleanly on 4.0.0 installations
and any subsequent updates should simply have additional sql
files in the schema_updates directory.

The sql files are named by version number and by database
type. An example of a file that can be ingested with postgres
and mysql is named "2.postgres.mysql.sql".  By convention,
we will keep the sql in source in this format, but package
the sql files separately (the mysql package will have a copy
of the earlier example file as just "2.mysql.sql").