[#2105] additional configuration options for service account details
This set of updates was requested for compatibility with installation
in network with a central naming authority (Kerberos, etc.) in which
service accounts need to be managed more specifically on each host.

The service account that owns and runs the iRODS server can now be
specified at the time of installation.

Service account creation is now handled by the setup_irods.sh
script and not the packaging.