[#2233] fix db update script for oracle upgrades
in testing upgrades on oracle the path where r_grid_configuration was
not present requires custom error handling for oracle due to differing
behavior from mysql and postgres.

also, irods_setup required different fallbacks when sqlplus failed as
it returns 0 when a connection cannot be made.

server_config was modified to add semicolons if none are present at
the end of a statement as these are required by oracle as well.