[#2105] removing overzealous environmental sanity check
Before allowing the service account to be specified at the time
of installation, this check was in place to ensure that whomever
was running the install scripts were doing it as the correct user.

This is no longer a safe assumption as an existing user with a
different $HOME is a perfectly legitimate use case.

Removing this check is harmless and allows for LDAP and
Kerberos-based configurations to exist as designed.