[#2485] add clean/stale/dirty replica behavior text
The iCAT stores a column named 'data_is_dirty' in R_DATA_MAIN.

The values in this column are indicative of whether the data on disk
has been updated.  In this 'in-flight' context, dirty means updated.

Otherwise, when data is at rest, 'dirty' is usually indicative of
being out of date, and a 'clean' replica would be a desirable copy
of the data.

This confusion is mostly limited to an implementation detail of the
iCAT itself.  The virtual layer provided by iRODS can be thought
to provide 'staleness' information via the '&' displayed before
'ils -l' and 'ils -L' shows the name of the Data Object.

The test associated with this iget behavior is here:
  https://github.com/irods/irods/blob/master/tests/pydevtest/resource_suite.py#L107