[#2789] generate new child list for create op
The selection of valid candidates to receive a replication
was done for the initiating operation, a READ in the case
of an icp.  Given a weighted passthrough child with a read
weight of 0.0, that child would not be in the list.  A new
list specifically for the reception of replicas was created
which will allow the children to vote directly on the
replication operation.