[#2854] makeTicket() was returning names w/ fewer than 15 chars
buf1 was modulo'd to find an offset into an array of printable
characters, but buf1 was of type (char[]), which behaves as signed
on x86, and e.g. (-15 % 62) is -15. The negative number was indexing
the array of printable characters and returning whatever happened to
be on the stack in that location.