[#2803] Check gethostbyname for TRY_AGAIN h_errno
In load tests, gethostbyname() was returning NULL in setSockAddr() on the client.
This caused the client to give up on the parallel transfer portal connection,
which led to select() timeouts in acceptSrvPortal() on the server.

gethostbyname() can set h_errno to TRY_AGAIN on NULL returns, however, which
lets us retry instead of aborting.