[#2970] do not create R_OBJT_FILESYSTEM_META in upgrade script
This was originally included to support the legacy 3.x codebase
and the possible 'run-as-root' option.  The code for this feature
has been removed from 4.x and this table is not necessary.

The Oracle flavor of the upgrade script was using an unsupported
'IF NOT EXISTS' syntax, which brought this to our attention.