[#3025] Use the canonical path of libirods_common.so
On cen7, /lib is a symlink to /usr/lib, and the runtime linker looks in
/lib before/instead of /usr/lib, so dladdr() returns /lib/libirods_common.so
as the path to libirods_common.so instead of /usr/lib/libirods_common.so.
This breaks the assumption that the path to the installation root is obtained
by calling remove_filename three times on the path to libirods_common.so.