[#3169] Prevent infinite recursion in irods-re when calling internal rule fails
If a rule in irods-re attempts to call another rule in irods-re with the wrong number
of arguments, the call will fail in execRule. If all the rule arguments were T_STRING,
the irods-re will then attempt to forward the rule call to the rule engine plugin framework
(REPF). The REPF will then ask irods-re if it has the rule in question, and the irods-re
will respond in the affirmative, because argument count checks are not part of the REPF
"rule_exists" interface. The REPF will then hand the rule call off to irods-re, restarting
the process.

Fix by not redirecting failed rule call to REPF if irods-re would tell the REPF that it does
have the rule.