[#3452] Avoid feeding untrusted input to the user's shell in igetwild
We extract a list of filenames from iquest, and then pass
them to iget one by one using xargs.

The complication is that filenames can contain pretty much any
character, and iquest doesn't have the option to separate output
records with the NULL byte. So we use the closest we can, which is /\n
(iquest always adds the \n). We then replace that /\n sequence with
NULL. That gets us most of the way there, but our sequence will still
finish "/\n", because sed doesn't see the final newline in its pattern
space. So we replace the final / with NULL and then use head to remove
the final character (the \n).

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>
Signed-off-by: Terrell Russell <terrellrussell@gmail.com>