[#3679] Fix dereference of null pointers in packStruct.cpp
extendPackedOutput is being used to populate outPtr in two places.
If this function fails, outPtr is explicitly set to null. This
change adds checks to ensure that extendPackedOuput did not fail
(and, therefore, that outPtr is not null) before dereferencing outPtr.
In the event of failure, SYS_MALLOC_ERR is returned.

(cherry-picked from SHA: 56fe86086a2229a3b0c9461ab565681afd400f8a)