[#3679] Fix dereference of null pointer in packStruct
When null-terminating the outPtr of extendPackedOutput, outPtr is
dereferenced in a static cast to a char pointer. It is possible for
extendPackedOutput to fail when allocating space for the packedOutput
buffer, and outPtr will be explicitly set to NULL.

This change adds a check for the error code indicating the failure
of extendPackedOutput before null-terminating outPtr. Failure now
prevents the "packed XML" log message  because the packedOutput
buffer will be NULL in the event of failure as well.