[#3679] Add null pointer checks in a few files
In l3Mkdir:
Fixed situation where dataObjInfo could be null when passed into
l3Mkdir. The intention was likely to only call l3Mkdir if specColl
was null or colClass == LINKED_COLL, not when dataObjInfo itself
is null.

In updateInEnv:
updateInEnv takes an Env pointer as argument, and the pointer can be
null. Since the entire function depends on env, ensure that it is
not null before proceeding with any operations.

In getDataObjInfoIncSpecColl:
This change adds a null check before dereferencing dataObjInfo or
checking specColl.

Also fixed constants going on the left in change with SHA:
eaab3c4549b954d2df701117d8a7b631449998fb