[#3701] Use scoped_threads to avoid mem leaks
When returning early, boost::threads were not being rejoined in
some places, which could potentially lead to memory leaks. This
change switches to using boost::scoped_threads which is an RAII
wrapper class for boost::thread. join() is called by default as
part of the dtor if the thread is joinable.