[#3619] move stop operation after cleanup, after chlClose()
The original commit to #3619 gave the new amqp audit plugin
the lifetime it needed to still be able to send messages before
the stack was destroyed.  That commit placed the stop_operations() before
cleanup(), which contains chlClose().  This meant that the dynamic
policy enforcement point for close() could not occur and resulted in
a stacktrace.

This commit has the effect of allowing that close() to occur cleanly,
but means the database is not available for the stop operation if
a rule engine plugin was written that tries to use it.  If such
a database call is desired, it will need to be a separate
database connection and hope that the consistency is enforced.  This
has not been tested.