[#3705] Fix iquest not returning DATA_RESC_HIER
If DATA_RESC_HIER is used in a GenQuery, the DATA_RESC_ID for the
hierarchy is used instead in the actual query and then replaced at the end
with DATA_RESC_HIER again. There is a bug where selecting DATA_RESC_HIER
as a column and using the DATA_RESC_HIER as a filter causes the replacement
of the ID at the end to fail. This change fixes this issue by not
allowing the inputs for the filter to override the saved position of
the column in the output.