[#3714] Fix GenQuery for resource hierarchies
DATA_RESC_HIER is no longer stored in the metadata table, so SQL cannot
be used to query resource hierarchies directly. As a result, using
DATA_RESC_HIER in the WHERE clause of a GenQuery can cause strange returns.
Using DATA_RESC_HIER in the WHERE clause of a GenQuery will now return a
result consistent with the rules below:

 - '=' with %: error
 - '=' without %: direct match for leaf
 - 'like' without %: direct match for leaf (like '=')
 - 'like' with %: consistent with 'like' for other columns