[#3265] Improve ichksum admin flag tests
Given a file that a rodsuser has permissions for and an admin user does
not, the new tests cover scenarios for running ichksum for all
combinations of the following:

1. Checksum already generated for file (or not)
2. Running as admin user (or rodsuser)
3. Running ichksum with admin flag (or not)

The admin flag should give permissions to the admin user regardless of
permissions with regards to the file. The rodsuser should not be able to
use the admin flag.