[#3436] Reduce log noise connection issues
Connection errors in sockComm give a lot of information that is not
particularly useful and makes the logs very noisy. Specifically,
printing the status in the message passed to rodsLogError is pointless
because rodsLogError prints the status anyway.

In the case of the "errno" message printed on connection being refused,
this change increases the level needed in debug logging in order to see
these messages. Fixed test_icommands_file_operations test to adjust for
this new behavior.