[#3436] Add retry capability to replication resc
This change adds a configurable retry mechanism to the replication resource.
The replication resources can be configured via 3 new settings in the
context string of the resource.

The retry applies when the replication from an iput or a replication
rebalance fails. The default value for each setting is 1, which means a
single retry after 1 second has passed.