[#3706] Make test_repl_num_policy count repls
test_repl_num_policy was trying to assert both the number of replicas
being made based on the num_repl context string and the destination
resources of those replicas. As the destination resource is
determined by voting (locality of reference, order in the catalog,
etc.), the test was technically non-deterministic and could fail on
occasion.

This change modifies the test to just count the number of replicas
being produced by the replication resource to ensure that the num_repl
policy is being enforced.