[#3904] Add test for repl redirect map regeneration
This change adds a test to ensure that multiple puts to a replication
resource across a single connection are regenerating the list of
children to replicate to between replications. The random resource
will demonstrate the differing replication targets by selecting a
random replication target after each replication.

The test recursively puts 50 files in a directory into an iRODS
collection, so this means that there is a 1/(2^49) chance that the
replication targets are the same across all the replications.

(adapted from SHA: 19d51466bd7fe7753f08d1c9f8b3484538b5436d)