[#3904] Generate new child list for each repl
The redirect map for the operation replication in the replication
resource was being re-used over a single connection via a check for
an existing property map value.

This change removes the check so that the redirect map is re-generated
for each replication.

(adapted from SHA: 1b758c216fffd4e8b0fd291c9ab85975d7e03c62)