[#3909] Compact replRedirect to one hier selection
replRedirect has two calls to the implementation function: the first
is to get the child list for replication targets on a create/write,
and the second is to get the child list for replication targets on
the actual operation to be performed. The second child list was
introduced to prevent a child from voting 0 on a consequential
create/write after voting 0 for the initial operation, as described
in irods/irods:2789.

This creates a situation where the selected hierarchy for a
create/write operation (chosen in first vote) could differ from the
selected hierarchy in the "initial" vote in replRedirect (which
could also be a create/write). This may lead to inconsistent behavior
as it relates to the child list.

This change consolidates the voting, swapping out the child list
property for the "write" version on create/write so that the child
list and selected hier are consistent across the operations.

Renamed some of the property map strings to reflect immutability.