[#3928] Error when root dir used with vault paths
When msi_update_unixfilesystem_resource_free_space() updates the free
space on a non-existent vault path (or one that boost::filesystem
interprets as non-existent), it checks the parent directories of the
resource vault path until it finds one that exists, even if it is the
root directory ('/' on UNIX systems).

Root directory should not be used as a vault path. An error is now
produced if the vault path is set to root directory. Additionally, the
root path should not be used in updating the free space on a resource.
This change causes msi_update_unixfilesystem_resource_free_space()
to log an error and exit if the path_to_stat becomes the root
directory because this gives misleading information about the free
space on a resource.