[#3927] Add enhanced logging server feature
Adds a class called enhanced_logger which holds a map of information
which can be accessed by logging utilities. The feature is enabled
through a new server_config.json string called `enhanced_logging`.
This is simply configured by configuring the specified setting to 0
for disable, and any non-zero value for enable. Defaults to 0.

Also added some additional logging in replication resource and the
unixfilesystem resource.