[#3932] Significantly enlarge objectMap hashtable to speed up rule caching.
At startup time and when rules are reloaded, parsed rule nodes are
copied to shared memory.  They are squeezed into a hashtable which
is much too small for large rulesets.  Let the size of the hashtable
depend on the size of the shared memory region (temporarily using
about half as much as the shared memory region for the hashtable).

(cherry-picked from SHA: 97d64b8ddeafd256cd8deeef0575d70b7d3179af)