[#3928] Add tests surrounding root dir and vaults
This change adds a test which moves the vault path of demoResc to a
non-existent directory under root directory and manually executes a
rule to call the msi_update_unixfilesystem_resource_free_space
microservice. The result should be an error along with an unchanged
free_space context value.

Also adds tests to ensure that using localhost as a host for a
resource produces a warning to the user. Adds tests to ensure that
root directory cannot be used as a vault path. Tests created for
both creating and modifying storage resources.

(cherry-picked from SHA: d5551ec2819a1a19b541681139593fe58e31d321)