[#47] Change iadmin to use obfEncodeByKey
obfEncodeByKeyV2 requires the challenge string from the server, which confounds decoding when not connected directly to a catalog provider. obfEncodeByKey works solely based on encoding the new password with another password (in all cases currently, specifically the password of the user calling it). This allows us to change the password even if we are not connected to a catalog provider server.