[#3981] Script to normalize unused R_DATA_MAIN columns
Adds a script to be run when updating past 4.2.4 which updates the
following columns in R_DATA_MAIN to contain "known values":
 - resc_name: EMPTY_RESC_NAME
 - resc_hier: EMPTY_RESC_HIER
 - resc_group_name: EMPTY_RESC_GROUP_NAME

Skips any rows with "invalid" resc_id column. A resc_id is considered
invalid if it does not exist in R_RESC_MAIN or it maps to a known
non-storage resource (i.e. coordinating resource).

Added get_database_type to database_interface for reusability.