[irods/irods#3991] Set spOption everywhere
spOption is the environment variable that the server checks to identify
a connected client. This variable should be set for all the icommands.
izonereport, iapitest, and iclienthints do not set this environment
variable. In the case of izonereport and iclienthints, the -h flag was
actually not supported at all, despite claiming to have support.