[#3991] Set spOption everywhere
spOption is the environment variable that the server checks to identify
a connected client. This variable should be set for all the icommands.
izonereport, iapitest, and iclienthints do not set this environment
variable. In the case of izonereport and iclienthints, the -h flag was
actually not supported at all, despite claiming to have support.

(cherry-picked from irods/irods_client_icommands SHA:
008b5343e086e452cbf62173d94fc4f8fbafd05a)