[#3998] Preserve remoteExecOut in rei->msParamArray
rei->msParamArray is used in smsi_remoteExec for remote execution
and clears the contents after completing. This was being done in
the past because state was being stored in the rule execution
environment and was used to populate rei.

This has changed as of ced5dd3121ea49fa017bd29e643ab7ac287d30eb.
ruleExecOut is no longer being stored in the rule execution
environment. As such, rei->msParamArray needs to be preserved.