[#3941] Prevent double rule exec in delay queue
Delayed execution queue should only process the maximum number of
concurrent processes (as specified in server_config.json) per batch.
Currently, the RE server executes one batch of queued processes and
then another batch of processes marked as "failed". This change fixes
a bug where successful processes were being processed in the failed
batch in addition to the first batch, causing twice the number of
processes to be executed.

There is no test associated with this change because the fix for the
linked issue masks the situation in which this bug is a problem.

(cherry-picked from SHA: 31a273f8363bffd8edb1b99ae7510d76cc33134e)