[#3906] Pack and unpack output parameters without value.
When creating a delayed rule, a copy of the REI and top-level arguments
is preserved.  Unfortunately, when one of the top level arguments is an
output parameter that has not been assigned a value yet, packing would
fail with the error SYS_PACK_INSTRUCT_FORMAT_ERR, preventing the delayed
rule from being started.

Handle the particular case that prevented successful packing of an
uninitialized parameter.