[#4066] Skip redirect in reg with size KW
When the size KW is provided in the call to rsPhyPathReg, there is
no need to stat the file. As such, there is no need to redirect to
the remote server at all. This also resolves an issue with stat'ing
0-length files even in the presence of the size KW.

(cherry-picked from SHA: f512362c99980e5a3a6188f1ad422e65ef3ca1a6)