[#4100] Add zone checking to rsGetHostForGet/Put
rsGetHostForGet and rsGetHostForPut need to make sure they are
connected to the correct zone before attempting any operation. If
a resource hierarchy is involved, the operation will fail because
the target could be on a remote zone with a resource heirarchy which
differs from the connected zone (or worse - one which doesn't differ
and data is pulled/put from/to the wrong zone).

(cherry-picked from SHA: eec41ba84bde5687f8fa83acbd377c41fcdfa821)