[#1805] wrote tests for PREP genquery iterator
Created the test module (test_prep_genquery_iterator.py) for the Python Rule
Engine Plugin's General Query iterator, new as of iRODS 4.2.5+ . Both the
row-at-a-time and paging-style iterators are put through their paces.

(See https://github.com/irods/irods_rule_engine_plugin_python for documentation
on the iterators and examples of how to use them.)

The running of these tests is dependent on the presence of /etc/irods/genquery.py;
if the module is not seen at that path (for example, if the python rule engine
plugin is not installed), all of the tests enumerated below are skipped.

ITERATOR TESTS

The following tests are currently enabled by default.

    1 test_multiples_256

      - use the paged iterator to query (as rods admin) from a set of 512 matching
        objects, verifying that the result set contains exactly that many rows.
        In this test, we also augment the python rulebase in core.py by the
        addition of a function that directly instantiates and uses the iterator.

    2 test_with_modified_rulebase_on_boundary

      - query 512 objects (as a non-privileged rods user) using several
        different page sizes; check each time that the correct number of pages
        are returned, with the expected "modulo page size" number of rows
        contained in the last page if applicable. A zero page-size parameter
        denotes a request to use the "yield one row at a time" style of python
        iterator in the test.

    3 test_with_modified_rulebase_off_boundary

      - Same as above but query off of a 256*N result-size boundary. This
        test arranges for a total of (512 + 7 * 8 * 9) rows to be returned from
        the query, so that flaws (if any) related to terminating on a page
        boundary are not significant.

SKIPPED TESTS

The tests described next are currently disabled by default, due to reasons
outlined later in this text:

    4 test_nested_paged_iterators

      - runs through paged + row + paged iterators of non-overlapping result
        sets, in three nested levels of looping, to produce a total 300*2*300
        results in the proper (ie monotonically increasing) order

    5 test_multiples_256_nested

      - the ambitious version of test # 1.   Here, 2*256*512 results are
        queried in row + row + paged iterator loops nested 3-deep.

There is an "auto-free-queries" feature which - if enabled - will cause these
last tests to be run, not skipped.  The tests are fully expected to pass and
not to cause ill effects.  However, note that if you choose to enable this
feature on an iRODS installation, error-free operation is not guaranteed!

See notes below (**) for details on enabling auto-free-queries in genquery.py.

______________ (optional reading! )

(**) Notes - Details and reasoning behind skipping some of the iterator tests

The skipped tests are resource-intensive.  Currently they are disabled due to
uncertain iRODS core semantics regarding when/if general query objects should
be closed out explicitly by the caller. This has yet to be fully investigated.
Tests and observations have borne out that the iRODS core sometimes will assume
questionable ownership of those objects and free them during the course of an
rsGenQuery call. Read on for more detail....

In the genquery.py iterators, as with all client code that leverages the bare
GenQuery microservice calls and expects the number of row results could exceed
the page size of MAX_SQL_ROWS (256), we monitor the value of continueInx returned
from msiGetMoreRows() between pages of results.  Ideally, a return value >= 1
for continueInx should unambiguously signal the availability of further result
pages, to be fetched by continuing to loop over calls to msiGetMoreRows(). In
contrast, a 0 value should signal it is time to quit the query loop and
(presumably) free resources using msiCloseGenquery().

But empirical evidence shows that the msiGetMoreRows() microservice can return
an error code of CAT_NO_ROWS_FOUND if its last valid iteration yields a full
page (which happens when the total number of rows returned by a query spanning
N pages is exactly N * MAX_SQL_ROWS).  At the time of this writing, a status
code of CAT_NO_ROWS_FOUND from msiGetMoreRows() seems to indicate:

    (1) the current results page is indeed the last one
    (2) no further calls to msiGetMoreRows() should be made
    (3) iRODS has internally freed objects and memory associated with the query

In the genquery.py module, by default a conservative resource-freeing policy is
in force:  specifically, we do not close the query ourselves.  This should not
be a problem for any short-running iRODS session, but it could lead to the
consumption of large amounts of memory if large numbers of queries are made
within the context, and lifetime, of one single iRODS agent process.

The iterator module's default choice of not freeing the query resources can be
altered if the application's memory requirements or time-of-average-run, or
both, require calling msiCloseGenQuery() at the appropriate times as dictated
by the heuristics just discussed.  The user can do so by assigning a boolean
True value to a the genquery.AUTO_FREE_QUERIES flag.
_______
(***) Notes to the IPUT fix

This delta also includes a correction of one of the iput tests which
showed indeterminate behavior during testing of the python iterator. The iput
test is this one:

  test_iput_options.Test_iPut_Options.test_iput_recursive_with_period__issue_2010

The problem was reasoned to stem from launching a recursive iput from a starting
point of the iRODS home directory itself, just one directory level above the Vault.
This can (but doesn't always) cause a recursion that refuses to bottom out, as
a file being "put" may potentially materialize in the Vault before iput has scanned
the directory node of the Vault subdirectory containing the new file copy.  The
symptom was that ever-longer paths would be referenced, eventually invoking the
error code: USER_PATH_EXCEEDS_MAX.

In the fix, we create a directory of a known structure just below the iRODS home
and then cd into it before performing the recursive iput from "./".