[#4099] Prevent file_modified on non-write ops
This change restricts when fileModified is called in rsModDataObjMeta.
This should prevent replication (in the case of replication or compound
hierarchies) on any non-write/create operation.

The replication resource has been modified such that the child replication
list is generated at the time of fileModified instead of in resolve
resource hierarchy. This ensures that the replication list is made
available to the machine doing the replication as the property map may
not be populated from machine to machine.

(cherry-picked from SHA: 5acfcd7d063018e864b5b62a39cccdfdae8119e4)