[#4147][#4148][#4179] Rule Engine Plugin Framework now supports continuation.
This is accomplished through the use of SYS_NOT_SUPPORTED and a new code
called RULE_ENGINE_CONTINUE.

If SYS_NOT_SUPPORTED or RULE_ENGINE_CONTINUE is returned to the REPF,
then the REPF will allow the next REP to respond to the given operation.
Else, it will exit immediately.

Introduces new Passthrough Rule Engine Plugin.