[#3776] Prevent double free in agent spawner
The main server agent spawner which handles connection requests can
fail when communicating with the agent factory. On failure, the
connection request is freed by the spawning thread and continues.
Later, procChildren comes to clean up the connection request queue
and also frees the connection request, which results in a double free.

This change puts a new message exchange between the main server and
the agent factory when establishing the connection. After the connection
is made, an acknowledgement is sent from the main server to the agent
factory. If anything other than the expected value (a success message)
is received, the agent factory will not fork an agent process.

The server process now consistently cleans up the open sockets and
associated socket files before returning from execAgent.