[#4240] Add query limit to query iterator
max_rows has changed to query_limit conceptually within the query
iterator. Now the user of query objects passes a limit for the
number of results that the iterator will traverse before stopping.
The default value of 0 means that there is no limit and the iterator
will continue until there are no more results. The iterator tracks
how many rows it has processed and when this value exceeds the limit
value, the iterator enters the end iteration state.

The query string is now copied into the query base implementation
object instead of just holding a reference in case the user has passed
a temporary when constructing the query object.