[#3998] Prevent reading of msParam with no type
In remote execution blocks, ruleExecOut is not always assigned a type.
This causes a strcmp on a null pointer to char in addMsParamToArray. This
change adds a check in smsi_remoteExec before copying around the ruleExecOut
param to ensure that its type is not null. A check is also done in
addMsParamToArray to ensure type is not null before attempting the strcmp.