[#4271] Add thread pool
Adds a light wrapper around a boost::asio::io_service and thread_group
to represent a thread_pool (boost::thread::thread_pool was introduced
in boost 1.66, which is newer than the version in use in 4-2-stable).
Each thread points to an io_service and a work object which can be
modified via free-standing functions irods::thread_pool::post and
irods::thread_pool::dispatch (again, modeled after boost::asio).

This is part of the iRODS dev package.