[#3782][#4250][#4251][#4266] Rework RE server
This change rewrites the Rule Execution server to use the new query
iterator, thread pool, and connection pool in order to speed up
processing of the delay queue.

The delay queue is now represented by a class which acts as a
thread-safe wrapper around a std::vector in which rules are held to
track what rules are set to be processed. This replaces the need to
change the status column for scheduled jobs in the rule exec table.
The RE server "enqueues" rules while worker threads "dequeue" the
rules after they have been processed.