[#4284] Add connection pool refresh timeout
Adds a refresh timeout mechanism to connections in a connection pool
such that verification of connections now includes a check for how
long the connection has been alive. If it exceeds the value stored in
irods_connection_pool_refresh_time_in_seconds (stored in iRODS env),
the connection is re-created.