[#4319] Test repl to all local leaf repl hier
When replicating a data object from a resource on the catalog provider
to a replication resource whose leaves are all co-resident on the same
catalog consumer, 0-length orphan files appear on the disk as well as
SYS_BAD_FILE_DESCRIPTOR error messages in the catalog consumer logs.
The size of the data object must exceed 1MB (1024*1024=1048576 bytes).

This test demonstrates this behavior.