[#4345] Fix tests failing after logging changes
This change fixes two tests which needed to be updated in order to
conform to the new logging system. spLogLevel now needs to be set to
at least the level of the corresponding logging category to see messages
of that level in the log. So, to see LOG_DEBUG messages from legacy
logging calls, spLogLevel needs to be set to at least 7 and the "legacy"
key in server_config under log_level needs to be at least "debug".

Also fixed a JSON tag for host name which was changed from the schema.

Also changed read/write permission for /var/log/irods so service account
can view the files.