[#4351] Add test for delay server resilience
Adds a test which kills a delayed rule execution agent to ensure that
the delay server survives. Adds a test which intentionally raises an
exception in the connection pool to ensure that the delay server survives.
This test, it should be noted, only produces an exception in the delay
server when the delay server does not exit on SIGPIPE, which is the
purpose of the first test mentioned above.

Also adds a test which ensures msiSegFault can only be run by admin users.