[#4351] Make delay server more resilient
Re-establish connection for deleting failed rules which may have been
caused by a broken pipe. Do not exit the delay server process on
SIGPIPE, which is sent when the agent executing a rule dies.

Wrap rule execution in thread task with a try-catch to handle exceptions
which may arise when rules are executing. It is possible that when removing
a rule from the delay queue after failure, an exception is thrown in the
delay server thread itself. This has been demonstrated by killing a rule
executing agent and interrupting the ODBC connection held by the delay
server when the failed rule is to be deleted from the queue.