[#4356] Make "pending SQL at cllDisconnect" true
A warning message is appearing in the logs stating that there is pending
SQL at cllDisconnect when in fact the number of pending statements is 0.
This change ensures that there are pending statements before attempting
to process or log them.

Also, sleep 30 seconds during setUp in test_icommands_file_operations.
This is to prevent log message suppression on Cen7 due to limitations
in the default configuration of journald.