[#4495] Chunk libarchive reads for extractions
The extract operation in structfile resource plugin reads in entire
files into a memory buffer without regard for size. If the file in
question overflows an int when its size in bytes is calculated,
libarchive will fail to read the file and extraction will do nothing.
This change iterates over the file using a block size and the buffer
pointer in use today. The block_size has been hard-coded to 1MB as
performance gains above this are negligible, and block sizes below
this only become increasingly worse as the value decreases.