[#4195] Don't let rsDataObjRead read empty replica
When a replica has a size of 0 in the catalog, the amount of data
which can be read from that replica should be 0. This short-circuits
rsDataObjRead by returning an empty bytesBuf if the data size in the
catalog for a given replica is 0, regardless of whether there is data
to be read on disk.