[#4943] Don't close() fds before calling dup2() on them
It is unnecessary to close() a fd before calling dup2() on it - dup2
will close the fd if necessary. Importantly, dup2 will do so
atomically, whereas the code here is racy - some other thread could
open() something inbetween the close() and dup2(), in which case that
other thread's FD will be overwritten.

This is a partial fix for #4943
Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>