[#4972] Support const struct in key_value_proxy
Adds template class for key_value_proxy to support const keyValPair_t
as well as enforcing the fact that the proxy only manages keyValPair_t's.

Several missing pieces were added to the iterator and handle classes to
flesh out the interface a little more.

get() was changed to return a pointer instead of a reference to the
underlying object in order to maintain const-ness of underlying struct.
Without this, a const keyValPair_t could be given to a proxy object and
returned as non-const.