[#3848] Populate dataId in dataObjInfo_t on create
This change populates dataObjInfo_t.dataId on create. The information
is available at the time of registration in the catalog but rsPhyPathReg
does not return it to the caller. So, creating a new replica has been
split into two paths: new replica for existing data object (which uses
rsRegReplica) and new replica for new data object (rsRegDataObj).

This change also makes create operations aware of the fact that it is a
replication or copy by the presence of a source L1 descriptor keyword.
This allows the destination replica to confirm its validity as well as
access information about itself based on the source replica.

Also rearranges rsPhyPathReg and boxes in some unused functions in the .hpp.