[#5150] Add replica_/data_object_proxy utilities
Adds a get_data_object_info function to the replica library. Returns an
irods::query::value_type which contains information for all replicas of
the specified data object. If a replica number is provided, the
information for only that replica will be returned.

Adds new factory methods for replica_proxy and data_object_proxy which
create new dataObjInfo_t structures and populate them with the catalog
operation mentioned above.

Adds a to_json free function to the replica namespace for converting
replica_proxy objects to a structured JSON format.